+++ quant/data_sources/providers/__init__.py
@@ -1,5 +1,6 @@
 from .akshare import AkShareProvider
 from .eastmoney import EastMoneyProvider
+from .fundamentals_legacy import LegacyFundamentalsProvider
 from .tencent import TencentProvider
 
-__all__ = ["AkShareProvider", "EastMoneyProvider", "TencentProvider"]
+__all__ = ["AkShareProvider", "EastMoneyProvider", "LegacyFundamentalsProvider", "TencentProvider"]